Move manipulation of binary distribution archives into separate module
One of the goals of this refactoring was to make the functions that
manipulate binary distribution archives easily re-usable outside of
pip-accel. While I was doing that (separating concerns) it became
obvious that these functions belonged in a separate module. This
moving around of responsibilities also necessitated moving the
configuration defaults into a separate module.

I also improved the tests by respecting the Python version inside
the virtual environment created by the tests, and added a test that
the installed executable actually works.