Check for /etc/debian_version
Since lsb-release does not tell you if the OS is a derivative, checking for /etc/debian_version (which is only there if the OS uses the debian base_files) tells us if the OS is a Debian based one.

This fixes dependency searches on, for example, Elementary OS.

The other option would be to check for /etc/os-release, but some other OSes like Fedora deprecated and removed this file some time ago, and it requires parsing the file, while this is much faster.