Read the Docs builds started failing, trying to get them running again
See e.g. https://readthedocs.org/builds/pip-accel/2533009/

This most likely has *something* to do with the release of pip-accel 0.23
which bumps the pip requirement to >= 6.0.8 (new versions of pip bundle
the `retrying' and `six' packages under pip._vendor). That doesn't explain
why a wrong (old) version of `six' would be bundled or loaded, but I
digress.

Why does Python packaging have to be such a mess, with creative and
fragile hacks everywhere?! :-(