Improve installed files tracking by forcing setuptools
When pip runs a setup.py script it forces the use of setuptools (instead
of distutils) which means *.egg-info metadata is stored in a directory.
In this directory pip creates an installed-files.txt file which enables
reliable uninstallation.

In pip-accel 0.25 (0bf3c2202d2) support for installed files tracking was
added, but pip-accel didn't force the use of setuptools in setup.py
scripts. Because of this every package whose setup.py script used
distutils would log the following warning during installation:

  Not tracking installed files (couldn't reliably determine *.egg-info directory)

With the changes in this commit all setup.py scripts will be forced to
use setuptools which stores *.egg-info metadata in a directory, enabling
tracking of installed files by pip and pip-accel, which in turn enables
removal of previously installed packages (pip uninstall).